auth-server: server: gas-sponsorship: Break out into separate module
