+++ auth/auth-server/src/server/gas_sponsorship/contract_interaction.rs
